이미지 추가 및 움직임 개선
